Support export/import plain keys in FIPS Mode.
The IBM Semeru FIPS mode currently leverages Mozilla's NSS library that is installed on RHEL platform. IBM Semeru talks to the NSS native library via the SunPKCS11 security provider which interfaces to NSS using the PKCS11 protocol. When in FIPS mode, the NSS library disables the export/import of cryptographic keys so that cryptographic keys are contained only within the scope of the NSS library. Attempting to extract cryptographic keys that were generated within NSS or import keys so NSS will perform computation on them (such as loading keys from a file) - will throw an exception.

This is a problem for some users such as Liberty, whose Lightweight Third Party Authentication (LTPA) feature requires importing and exporting cryptographic keys. This PR allows a user to export and import cryptographic keys by leveraging the PKCS11 wrapping/unwrapping feature which does allows to indrectly extract/insert keys.

The idea of this PR is to intercept at the point at which a key is imported/exported under regular non-FIPS mode, detect it is in FIPS mode and perform unwrap/wrap to move the key across the PKCS11 module boundary. For importing a key, this is done in the CreateObject call - for exporting a key this is done in the P11Key creation code. The reason we need to modify the pkcs11 security provider at a fairly low level is that we require access to the underlying pkcs11 token in order to invoke the unwrap/wrap calls.

Signed-off-by: Alon Shalev Housfater <alonsh@ca.ibm.com>