Get CKA_VALUE from derived secretkey in FIPS mode
There is an issue in FIPS mode while
generating a secretkey in keyagreement
process. Currently we can not get the
CKA_VALUE from a derived secretkey in
FIPS mode during the keyagreement.

Our solution is using a Java code to
replace the native C_GetAttributeValue
function to get the CKA_VALUE from the
derived secretkey.